+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSFTNWCJW9RVKK72NG7ANJCF"
+        "01KSFYHD383GAR720SG0AR0MCR"
     )
     r"""The unique identifier of the data source"""