+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSFTNWCPY41YFFQPC216FZH9"
+        "01KSFYHD3DEYPD4T8CB6N1YHZ0"
     )
     r"""The unique identifier of the data source"""