+++ packages/orq-rc/RELEASES.md
@@ -8179,3 +8179,13 @@ Based on:
 - [python v4.10.0-rc.32] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.32 - packages/orq-rc
+
+## 2026-05-25 16:15:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.765.0 (2.886.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.33] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.33] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.33 - packages/orq-rc